course purhcase endpoint

+++ controllers/courseDelete.js
@@ -3,6 +3,17 @@ const { Course } = require("../db");
 async function courseDelete(req, res) {
   const courseId = req.params.courseId;
 
+  const foundCourse = await Course.findOne({
+    _id: courseId,
+    creatorId: req.id,
+  });
+  if (!foundCourse) {
+    return res.status(401).json({
+      success: false,
+      message: "You cannot delete someone else's course",
+    });
+  }
+
   try {
     const updatedCourse = await Course.findByIdAndDelete(courseId);
